requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.3
psycopg2==2.9.3
python-decouple==3.6
sqlparse==0.4.2
